Migrates remaining projects to .NET Core or xproj (#635)
* Adding xproj for all csprojs

* Adding project.json templates to each project

* Change version to 99.0.0-dev to stop restoring packages from myget feed

* Migrate EasyNetQ.Hosepipe.SetupActions to xproj

* Adding NuGet v3 props/targets as ignoreable

* Add EasyNetQ xprojs to solution

* Migrate EasyNetQ.Hosepipe to .NET Core

* Migrate EasyNetQ.LogReader to xproj

* Migrate EasyNetQ.Trace to .NET Core

* Migrate EasyNetQ.Scheduler.Mongo.Core to .NET Core

* Migrate EasyNetQ.Scheduler.Mongo.Tests to .NET Core

* Migrate Source/EasyNetQ.Scheduler.Mongo to xproj

* Migrate EasyNetQ.Scheduler.Tests to xproj

* EasyNetQ.Serilog: Update Serilog verison

* Adding runtimes to EasyNetQ.Tests

+++ Source/EasyNetQ.Hosepipe.Tests/FileMessageWriterTests.cs
@@ -10,7 +10,7 @@ namespace EasyNetQ.Hosepipe.Tests
     [TestFixture]
     public class FileMessageWriterTests
     {
-        private const string tempDirectory = @"C:\temp\MessageOutput";
+        private readonly string tempDirectory = Path.Combine(Path.GetTempPath(), @"MessageOutput");
 
         [SetUp]
         public void SetUp() {}
@@ -19,9 +19,17 @@ namespace EasyNetQ.Hosepipe.Tests
         public void WriteSomeFiles()
         {
             var directory = new DirectoryInfo(tempDirectory);
-            foreach (var file in directory.EnumerateFiles())
+
+            if (!directory.Exists)
             {
-                file.Delete();
+                directory.Create();
+            }
+            else
+            {
+                foreach (var file in directory.EnumerateFiles())
+                {
+                    file.Delete();
+                }
             }
 
             var properties = new MessageProperties();
